refactor: split pdrutils.py into pdrtpy/utils/ subpackage (#211)
* remove unused code

* refactor: split pdrutils.py into pdrtpy/utils/ subpackage

Move code from pdrutils.py into five focused modules under pdrtpy/utils/:
- units.py: radiation field unit definitions and conversions
- paths.py: directory/path helpers and table I/O
- fits.py: FITS keyword utilities
- wcs.py: WCS and image array helpers
- helpers.py: general helpers (warn, is_image, H2 trim utilities, etc.)

pdrutils.py is retained as a backward-compatibility shim for external
code using 'import pdrtpy.pdrutils as utils'. extinct() and deextinct()
are removed (were no-op stubs).

All internal package imports updated to use pdrtpy.utils directly.
All 228 tests pass.

* bump for pdrutils refactor

* refactor: update test and table code to import pdrtpy.utils directly
