+++ pdrtpy/plot/test/test_plot_comprehensive.py
@@ -6,7 +6,7 @@ matplotlib.use("Agg")
 
 import astropy.units as u
 import numpy as np
-import pdrtpy.pdrutils as utils
+import pdrtpy.utils as utils
 import pytest
 from astropy.nddata import StdDevUncertainty
 from pdrtpy.measurement import Measurement
