+++ pdrtpy/test/test_measurement.py
@@ -3,7 +3,7 @@ import os
 
 import astropy.units as u
 import numpy as np
-import pdrtpy.pdrutils as utils
+import pdrtpy.utils as utils
 import pytest
 from astropy.nddata import StdDevUncertainty
 from pdrtpy.measurement import Measurement
